requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-filter==21.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
phonenumbers==8.12.50
pytz==2022.1
sqlparse==0.4.2
